requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pandas
numpy
xgboost
ipykernel
scikit-learn
matplotlib
seaborn
kagglehub
apache-airflow
apache-airflow-providers-amazon
# lightgbm
xgboost
mlflow
optuna
airflow-provider-mlflow
mlflow-skinny
astro-sdk-python